Fixed type confliect in ft_strrepeat

+++ src/ft_strrepeat.c
@@ -6,7 +6,7 @@
 /*   By: bbrassar <bbrassar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/30 18:09:42 by bbrassar          #+#    #+#             */
-/*   Updated: 2021/06/13 15:31:57 by bbrassar         ###   ########.fr       */
+/*   Updated: 2021/06/15 16:28:19 by bbrassar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrepeat(char const *s, int times)
 {
-	size_t	i;
+	int		i;
 	size_t	len;
 	char	*str;
